move tests for submissions to a folder
