+++ src/api/proposal-submission/api-v1-pdf-user-download-test.ts
@@ -1,13 +1,19 @@
 import { check, sleep } from 'k6';
 import http from 'k6/http';
-import { getEnvironmentConfigurations } from '../support/configurations';
+
 import exec from 'k6/execution';
-import { UserLogin } from '../utils/sharedType';
-import { randomIntBetween } from '../utils/helperFunctions';
-import { generateBearerToken, getAsyncClientApi } from '../support/graphql';
-import { Proposal } from '../graphql/support/proposal';
+
+
+
+
 import { Counter, Trend } from 'k6/metrics';
-import { User } from '../graphql/support/user';
+import { generateBearerToken, getAsyncClientApi } from '../../support/graphql';
+import { getEnvironmentConfigurations } from '../../support/configurations';
+import { Proposal } from '../../graphql/support/proposal';
+import { UserLogin } from '../../utils/sharedType';
+import { randomIntBetween } from '../../utils/helperFunctions';
+import { User } from '../../graphql/support/user';
+
 
 const environmentConfig = getEnvironmentConfigurations();
 const testVus=+__ENV.K6_PS_VUS || 50;
